test: move test shared code to common directory
