+++ test/asset/fixtures.ts
@@ -1,109 +1,5 @@
-import {Event} from '@ethersproject/contracts';
-import {
-  ethers,
-  deployments,
-  getUnnamedAccounts,
-  getNamedAccounts,
-} from 'hardhat';
-
-import {waitFor, setupUsers} from '../utils';
+import {deployments} from 'hardhat';
+import {assetFixtures} from '../common/fixtures/asset';
 // import asset_regenerate_and_distribute from '../../setup/asset_regenerate_and_distribute';
 
-export const setupAsset = deployments.createFixture(async function () {
-  // await asset_regenerate_and_distribute(hre);
-  const unnamedAccounts = await getUnnamedAccounts();
-  const otherAccounts = [...unnamedAccounts];
-  const minter = otherAccounts[0];
-  otherAccounts.splice(0, 1);
-
-  const {assetBouncerAdmin} = await getNamedAccounts();
-
-  const assetContractAsBouncerAdmin = await ethers.getContract(
-    'Asset',
-    assetBouncerAdmin
-  );
-  await waitFor(assetContractAsBouncerAdmin.setBouncer(minter, true));
-  const Asset = await ethers.getContract('Asset', minter);
-  const predicate = await ethers.getContract('ERC1155_PREDICATE');
-  const TRUSTED_FORWARDER = await deployments.get('TRUSTED_FORWARDER');
-  const trustedForwarder = await ethers.getContractAt(
-    'TestMetaTxForwarder',
-    TRUSTED_FORWARDER.address
-  );
-
-  // Set predicate Asset
-  try {
-    await waitFor(predicate.setAsset(Asset.address));
-  } catch (e) {
-    console.log(e);
-  }
-
-  let id = 0;
-  const ipfsHashString =
-    '0x78b9f42c22c3c8b260b781578da3151e8200c741c6b7437bafaff5a9df9b403e';
-
-  async function mintAsset(to: string, value: number, hash = ipfsHashString) {
-    // Asset to be minted
-    const creator = to;
-    const packId = ++id;
-    const supply = value;
-    const rarity = 0;
-    const owner = to;
-    const data = '0x';
-
-    let receipt;
-    try {
-      receipt = await waitFor(
-        Asset.mint(creator, packId, hash, supply, rarity, owner, data)
-      );
-    } catch (e) {
-      console.log(e);
-    }
-
-    const event = receipt?.events?.filter(
-      (event: Event) => event.event === 'TransferSingle'
-    )[0];
-    if (!event) {
-      throw new Error('no TransferSingle event after mint single');
-    }
-    return event.args?.id;
-  }
-
-  async function mintMultiple(
-    to: string,
-    supplies: number[],
-    hash = ipfsHashString
-  ): Promise<number[]> {
-    const creator = to;
-    const packId = ++id;
-    const rarity = 0;
-    const owner = to;
-    const data = '0x';
-
-    const tx = await Asset.mintMultiple(
-      creator,
-      packId,
-      hash,
-      supplies,
-      rarity,
-      owner,
-      data
-    );
-    const receipt = await tx.wait();
-
-    return receipt.events.find(
-      (v: Event) => v.event === 'TransferBatch'
-    ).args[3];
-  }
-
-  const users = await setupUsers(otherAccounts, {Asset});
-
-  return {
-    Asset,
-    users,
-    mintAsset,
-    mintMultiple,
-    trustedForwarder,
-    predicate,
-  };
-});
+export const setupAsset = deployments.createFixture(assetFixtures);
