+++ test/common/fixtures/erc20BasicApproveExtension.ts
@@ -1,0 +1,68 @@
+import {
+  deployments,
+  ethers,
+  getNamedAccounts,
+  getUnnamedAccounts,
+} from 'hardhat';
+import {Contract} from 'ethers';
+import {setupUser, waitFor} from '../../utils';
+
+// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+export const erc20BasicApproveExtensionFixtures = async function () {
+  await deployments.fixture([
+    'Land',
+    'Sand',
+    'LandPreSale_5',
+    'ERC20_PREDICATE',
+  ]);
+  const accounts = await getNamedAccounts();
+  const others = await getUnnamedAccounts();
+  const user0 = others[0];
+  const user1 = others[1];
+  const landSaleName = 'LandPreSale_5_16';
+  const deploymentData = await deployments.get(landSaleName);
+  const lands = deploymentData.linkedData;
+  const landContract: Contract = await ethers.getContract('Land');
+  const landSaleContract: Contract = await ethers.getContract(landSaleName);
+  const sandContract = await ethers.getContract('Sand');
+  const sandContractAsUser0 = await sandContract.connect(
+    ethers.provider.getSigner(user0)
+  );
+  await deployments.deploy('EmptyContract', {
+    from: accounts.deployer,
+    args: [],
+  });
+  const emptyContract: Contract = await ethers.getContract('EmptyContract');
+  await deployments.deploy('MockERC20BasicApprovalTarget', {
+    from: accounts.deployer,
+    args: [],
+  });
+  const mockERC20BasicApprovalTarget: Contract = await ethers.getContract(
+    'MockERC20BasicApprovalTarget'
+  );
+  const predicate = await ethers.getContract('ERC20_PREDICATE');
+  // Set predicate Sand
+  try {
+    await waitFor(predicate.setToken(sandContract.address));
+  } catch (e) {
+    console.log(e);
+  }
+  const sandBeneficiary = await setupUser(accounts.sandBeneficiary, {
+    sandContract,
+  });
+
+  return {
+    landContract,
+    lands,
+    landSaleContract,
+    mockERC20BasicApprovalTarget,
+    emptyContract,
+    sandContractAsUser0,
+    sandContract,
+    sandAdmin: accounts.sandAdmin,
+    sandBeneficiary,
+    user0,
+    user1,
+    predicate,
+  };
+};
